remove xmlExport function to modules
